fix(TDI-42230):Bug_TDI_26493_tFileCopy_LargerFile failed java 11 * move build from ant to maven
